Move wx-specific viewer code to ui/wx.
